pyproject.toml
[project]
name = "hsmw-tutor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.0",
    "gin-config>=0.5.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.1.0",
    "pymupdf>=1.27.1",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
]
